+++ source/FlexTable/FlexTable.example.js
@@ -1,13 +1,17 @@
 /** @flow */
-import React, { Component } from 'react'
+import Immutable from 'immutable'
+import React, { Component, PropTypes } from 'react'
 import { ContentBox, ContentBoxHeader, ContentBoxParagraph } from '../demo/ContentBox'
 import { LabeledInput, InputRow } from '../demo/LabeledInput'
 import FlexColumn from './FlexColumn'
 import FlexTable, { SortDirection } from './FlexTable'
-import Immutable from 'immutable'
 import styles from './FlexTable.example.css'
 
 export default class TableExample extends Component {
+  static propTypes = {
+    list: PropTypes.instanceOf(Immutable.List).isRequired
+  }
+
   constructor (props, context) {
     super(props, context)
 
@@ -20,17 +24,6 @@ export default class TableExample extends Component {
       headerHeight: 30,
       rowHeight: 40
     }
-
-    // HACK :)
-    const list = []
-    for (var i = 0; i < 1000; i++) {
-      list.push({
-        id: i,
-        name: `Item ${i + 1}`,
-        random: loremIpsum[i % loremIpsum.length]
-      })
-    }
-    this._list = Immutable.List(list)
 
     this._noRowsRenderer = this._noRowsRenderer.bind(this)
     this._onRowsCountChange = this._onRowsCountChange.bind(this)
@@ -49,8 +42,9 @@ export default class TableExample extends Component {
       rowHeight
     } = this.state
 
-    const list = this._list
-      .sortBy(index => this._list.getIn([index, sortBy]))
+    const { list } = this.props
+    const filteredList = list
+      .sortBy(index => list.getIn([index, sortBy]))
       .update(list =>
         sortDirection === SortDirection.DESC
           ? list.reverse()
@@ -58,11 +52,11 @@ export default class TableExample extends Component {
       )
 
     function rowGetter (index) {
-      return list.get(index)
+      return filteredList.get(index)
     }
 
     return (
-      <ContentBox className={styles.FlexTableExample}>
+      <ContentBox {...this.props}>
         <ContentBoxHeader
           text='FlexTable'
           sourceLink='https://github.com/bvaughn/react-virtualized/blob/master/source/FlexTable/FlexTable.example.js'
@@ -113,7 +107,6 @@ export default class TableExample extends Component {
         <FlexTable
           ref='Table'
           className={styles.FlexTable}
-          width={430}
           headerClassName={styles.headerColumn}
           headerHeight={headerHeight}
           height={height}
@@ -125,6 +118,7 @@ export default class TableExample extends Component {
           sort={this._sort}
           sortBy={sortBy}
           sortDirection={sortDirection}
+          width={430}
         >
           <FlexColumn
             label='Index'
@@ -165,7 +159,7 @@ export default class TableExample extends Component {
 
   _onRowsCountChange (event) {
     let rowsCount = parseInt(event.target.value, 10) || 0
-    rowsCount = Math.max(0, Math.min(this._list.size, rowsCount))
+    rowsCount = Math.max(0, Math.min(this.props.list.size, rowsCount))
 
     this.setState({ rowsCount })
   }
@@ -195,25 +189,3 @@ export default class TableExample extends Component {
     this.setState({ sortBy, sortDirection })
   }
 }
-
-const loremIpsum = [
-  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-  'Phasellus vulputate odio commodo tortor sodales, et vehicula ipsum viverra.',
-  'Cras tincidunt nisi in urna molestie varius.',
-  'Curabitur ac enim dictum arcu varius fermentum vel sodales dui.',
-  'Ut tristique augue at congue molestie.',
-  'Cras eget enim nec odio feugiat tristique eu quis ante.',
-  'Phasellus eget enim vitae nunc luctus sodales a eu erat.',
-  'Nulla bibendum quam id velit blandit dictum.',
-  'Donec dignissim mi ac libero feugiat, vitae lacinia odio viverra.',
-  'Praesent vel lectus venenatis, elementum mauris vitae, ullamcorper nulla.',
-  'Quisque sollicitudin nulla nec tellus feugiat hendrerit.',
-  'Vestibulum a eros accumsan, lacinia eros non, pretium diam.',
-  'Donec ornare felis et dui hendrerit, eget bibendum nibh interdum.',
-  'Donec nec diam vel tellus egestas lobortis.',
-  'Sed ornare nisl sit amet dolor pellentesque, eu fermentum leo interdum.',
-  'Sed eget mauris condimentum, molestie justo eu, feugiat felis.',
-  'Sed luctus justo vitae nibh bibendum blandit.',
-  'Nulla ac eros vestibulum, mollis ante eu, rutrum nulla.',
-  'Sed cursus magna ut vehicula rutrum.'
-]