+++ source/VirtualScroll/VirtualScroll.example.js
@@ -1,29 +1,23 @@
 /**
  * @flow
  */
-import React, { Component } from 'react'
+import Immutable from 'immutable'
+import React, { Component, PropTypes } from 'react'
+import styles from './VirtualScroll.example.css'
+import VirtualScroll from './VirtualScroll'
 import { ContentBox, ContentBoxHeader, ContentBoxParagraph } from '../demo/ContentBox'
 import { LabeledInput, InputRow } from '../demo/LabeledInput'
-import VirtualScroll from './VirtualScroll'
-import styles from './VirtualScroll.example.css'
 
 export default class VirtualScrollExample extends Component {
+  static propTypes = {
+    list: PropTypes.instanceOf(Immutable.List).isRequired
+  }
+
   constructor (props) {
     super(props)
 
-    // HACK :)
-    this._list = []
-    for (var i = 0; i < 1000; i++) {
-      this._list.push({
-        index: i,
-        color: BADGE_COLORS[i % BADGE_COLORS.length],
-        name: NAMES[i % NAMES.length],
-        height: ROW_HEIGHTS[Math.floor(Math.random() * ROW_HEIGHTS.length)]
-      })
-    }
-
     this.state = {
-      rowsCount: this._list.length,
+      rowsCount: props.list.size,
       scrollToIndex: undefined,
       useDynamicRowHeight: false,
       virtualScrollHeight: 300,
@@ -48,7 +42,7 @@ export default class VirtualScrollExample extends Component {
     } = this.state
 
     return (
-      <ContentBox className={styles.VirtualScrollExample}>
+      <ContentBox {...this.props}>
         <ContentBoxHeader
           text='VirtualScroll'
           sourceLink='https://github.com/bvaughn/react-virtualized/blob/master/source/VirtualScroll/VirtualScroll.example.js'
@@ -104,7 +98,6 @@ export default class VirtualScrollExample extends Component {
         <VirtualScroll
           ref='VirtualScroll'
           className={styles.VirtualScroll}
-          width={310}
           height={virtualScrollHeight}
           noRowsRenderer={this._noRowsRenderer}
           rowsCount={rowsCount}
@@ -117,7 +110,8 @@ export default class VirtualScrollExample extends Component {
   }
 
   _getRowHeight (index) {
-    return this._list[index].height
+    const { list } = this.props
+    return list.get(index).height
   }
 
   _noRowsRenderer () {
@@ -129,8 +123,10 @@ export default class VirtualScrollExample extends Component {
   }
 
   _onRowsCountChange (event) {
+    const { list } = this.props
+
     let rowsCount = parseInt(event.target.value, 10) || 0
-    rowsCount = Math.max(0, Math.min(this._list.length, rowsCount))
+    rowsCount = Math.max(0, Math.min(list.size, rowsCount))
 
     this.setState({ rowsCount })
   }
@@ -147,17 +143,18 @@ export default class VirtualScrollExample extends Component {
   }
 
   _rowRenderer (index) {
+    const { list } = this.props
     const { useDynamicRowHeight, virtualScrollRowHeight } = this.state
 
-    const datum = this._list[index]
+    const datum = list.get(index)
     const height = useDynamicRowHeight
-      ? datum.height
+      ? datum.size
       : virtualScrollRowHeight
 
     let additionalContent
 
     if (useDynamicRowHeight) {
-      switch (datum.height) {
+      switch (datum.size) {
         case 75:
           additionalContent = <div>It is medium-sized.</div>
           break
@@ -192,7 +189,7 @@ export default class VirtualScrollExample extends Component {
         </div>
         {useDynamicRowHeight &&
           <span className={styles.height}>
-            {datum.height}px
+            {datum.size}px
           </span>
         }
       </div>
@@ -206,7 +203,3 @@ export default class VirtualScrollExample extends Component {
     })
   }
 }
-
-const BADGE_COLORS = ['#f44336', '#3f51b5', '#4caf50', '#ff9800', '#2196f3', '#374046', '#cddc39', '#2196f3', '#9c27b0', '#ffc107', '#009688', '#673ab7', '#ffeb3b', '#cddc39', '#795548']
-const NAMES = ['Peter Brimer', 'Tera Gaona', 'Kandy Liston', 'Lonna Wrede', 'Kristie Yard', 'Raul Host', 'Yukiko Binger', 'Velvet Natera', 'Donette Ponton', 'Loraine Grim', 'Shyla Mable', 'Marhta Sing', 'Alene Munden', 'Holley Pagel', 'Randell Tolman', 'Wilfred Juneau', 'Naida Madson', 'Marine Amison', 'Glinda Palazzo', 'Lupe Island', 'Cordelia Trotta', 'Samara Berrier', 'Era Stepp', 'Malka Spradlin', 'Edward Haner', 'Clemencia Feather', 'Loretta Rasnake', 'Dana Hasbrouck', 'Sanda Nery', 'Soo Reiling', 'Apolonia Volk', 'Liliana Cacho', 'Angel Couchman', 'Yvonne Adam', 'Jonas Curci', 'Tran Cesar', 'Buddy Panos', 'Rosita Ells', 'Rosalind Tavares', 'Renae Keehn', 'Deandrea Bester', 'Kelvin Lemmon', 'Guadalupe Mccullar', 'Zelma Mayers', 'Laurel Stcyr', 'Edyth Everette', 'Marylin Shevlin', 'Hsiu Blackwelder', 'Mark Ferguson', 'Winford Noggle', 'Shizuko Gilchrist', 'Roslyn Cress', 'Nilsa Lesniak', 'Agustin Grant', 'Earlie Jester', 'Libby Daigle', 'Shanna Maloy', 'Brendan Wilken', 'Windy Knittel', 'Alice Curren', 'Eden Lumsden', 'Klara Morfin', 'Sherryl Noack', 'Gala Munsey', 'Stephani Frew', 'Twana Anthony', 'Mauro Matlock', 'Claudie Meisner', 'Adrienne Petrarca', 'Pearlene Shurtleff', 'Rachelle Piro', 'Louis Cocco', 'Susann Mcsweeney', 'Mandi Kempker', 'Ola Moller', 'Leif Mcgahan', 'Tisha Wurster', 'Hector Pinkett', 'Benita Jemison', 'Kaley Findley', 'Jim Torkelson', 'Freda Okafor', 'Rafaela Markert', 'Stasia Carwile', 'Evia Kahler', 'Rocky Almon', 'Sonja Beals', 'Dee Fomby', 'Damon Eatman', 'Alma Grieve', 'Linsey Bollig', 'Stefan Cloninger', 'Giovanna Blind', 'Myrtis Remy', 'Marguerita Dostal', 'Junior Baranowski', 'Allene Seto', 'Margery Caves', 'Nelly Moudy', 'Felix Sailer']
-const ROW_HEIGHTS = [50, 75, 100]